Introduce abstract HasError function for RangeValidator through which derived classes determine range compiance
