--- src/FluentValidation/Validators/RangeValidatorBase.cs
@@ -1,71 +1,0 @@
-﻿#region License
-// Copyright (c) .NET Foundation and contributors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// The latest version of this file can be found at https://github.com/FluentValidation/FluentValidation
-#endregion
-
-namespace FluentValidation.Validators {
-	using System;
-	using System.Collections.Generic;
-
-	public abstract class RangeValidatorBase<T, TProperty> : PropertyValidator<T, TProperty>, IBetweenValidator {
-
-		readonly IComparer<TProperty> _explicitComparer;
-
-		public RangeValidatorBase(TProperty from, TProperty to, IComparer<TProperty> comparer) {
-			To = to;
-			From = from;
-
-			_explicitComparer = comparer;
-
-			if (comparer.Compare(to, from) == -1) {
-				throw new ArgumentOutOfRangeException(nameof(to), "To should be larger than from.");
-			}
-		}
-
-		public TProperty From { get; }
-		public TProperty To { get; }
-
-		object IBetweenValidator.From => From;
-		object IBetweenValidator.To => To;
-
-		public abstract bool Validate(TProperty value);
-
-		public override bool IsValid(ValidationContext<T> context, TProperty value) {
-			// If the value is null then we abort and assume success.
-			// This should not be a failure condition - only a NotNull/NotEmpty should cause a null to fail.
-			if (value == null) return true;
-
-			if (Validate(value)) {
-
-				context.MessageFormatter
-					.AppendArgument("From", From)
-					.AppendArgument("To", To)
-					.AppendArgument("Value", value);
-
-				return false;
-			}
-			return true;
-		}
-
-		protected int Compare(TProperty a, TProperty b) {
-			return _explicitComparer.Compare(a, b);
-		}
-
-		protected override string GetDefaultMessageTemplate(string errorCode) {
-			return Localized(errorCode, Name);
-		}
-	}
-}
